backend/requirements.txt
Flask==2.3.3
face-recognition==1.3.0
opencv-python==4.8.0.76
numpy==1.24.3
Pillow==10.0.0
flask-sqlalchemy==3.0.5